requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
pillow
mlflow